# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI-Юрист Казахстан API
  version: 0.1.0
  description: REST contract for AI-Юрист Казахстан.
servers:
  - url: /api/v1
paths:
...
  /admin/legal-categories:
    get:
      operationId: adminListLegalCategoryTree
      summary: Admin list legal categories as a review tree
      parameters:
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      responses:
        "200":
          description: Admin legal category tree
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/LegalCategory"
                    - type: object
                      properties:
                        children:
                          type: array
                          items:
                            $ref: "#/components/schemas/LegalCategory"
  /admin/legal-categories/change-requests:
    get:
      operationId: adminListLegalCategoryChangeRequests
      summary: List legal category change requests
      parameters:
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      responses:
        "200":
          description: Legal category change requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalCategoryChangeRequest"
    post:
      operationId: adminCreateLegalCategoryChangeRequest
      summary: Create a safe legal category change request
      parameters:
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLegalCategoryChangeRequest"
      responses:
        "201":
          description: Created pending change request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCategoryChangeRequest"
  /admin/legal-categories/change-requests/{id}:
    get:
      operationId: adminGetLegalCategoryChangeRequest
      summary: Get legal category change request details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      responses:
        "200":
          description: Legal category change request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCategoryChangeRequest"
  /admin/legal-categories/change-requests/{id}/approve:
    post:
      operationId: adminApproveLegalCategoryChangeRequest
      summary: Approve and apply a legal category change request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      responses:
        "201":
          description: Approved change request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCategoryChangeRequest"
  /admin/legal-categories/change-requests/{id}/reject:
    post:
      operationId: adminRejectLegalCategoryChangeRequest
      summary: Reject a legal category change request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "201":
          description: Rejected change request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCategoryChangeRequest"
  /admin/classifications/review-queue:
    get:
      operationId: adminClassificationReviewQueue
      summary: Admin review queue for unconfirmed, high-risk or incomplete classifications
      parameters:
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      responses:
        "200":
          description: Classification records requiring admin or expert review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseClassificationRecord"
  /admin/classifications/{id}/confirm:
    post:
      operationId: adminConfirmClassification
      summary: Admin confirms a classification after review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      responses:
        "201":
          description: Confirmed classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /admin/classifications/{id}/override:
    post:
      operationId: adminOverrideClassification
      summary: Admin or expert corrects a classification and records expert feedback
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-role
          required: true
          schema:
            type: string
            enum: [admin, superadmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subcategoryCode]
              properties:
                subcategoryCode:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Expert-corrected classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /cases:
    get:
      operationId: listCases
      summary: List cases owned by current user
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Legal cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalCase"
    post:
      operationId: createCase
      summary: Create a legal case from confirmed text
      parameters:
        - in: header
          name: idempotency-key
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCase"
  /case-categories:
    get:
      operationId: listCaseCategories
      summary: List Kazakhstan case classification categories
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case classification taxonomy with criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseCategory"
  /legal-categories:
    get:
      operationId: listLegalCategories
      summary: List database-backed Kazakhstan legal categories
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Legal category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalCategory"
  /legal-categories/tree:
    get:
      operationId: listLegalCategoryTree
      summary: List legal categories as a tree
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Legal category tree
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/LegalCategory"
                    - type: object
                      properties:
                        children:
                          type: array
                          items:
                            $ref: "#/components/schemas/LegalCategory"
  /cases/{caseId}:
    get:
      operationId: getCase
      summary: Get case by id
      parameters:
        - in: path
          name: caseId
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Legal case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCase"
  /cases/{caseId}/classification:
    get:
      operationId: getCaseClassification
      summary: Get latest classification for a case
      parameters:
        - in: path
          name: caseId
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /ai/classifications:
    post:
      operationId: createClassification
      summary: Classify confirmed voice or text into an allowed KZ dispute category
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassificationRequest"
      responses:
        "201":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /ai/classifications/{id}:
    get:
      operationId: getClassification
      summary: Get classification by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /ai/classifications/{id}/clarifications:
    post:
      operationId: answerClassificationClarifications
      summary: Save answers to category clarification questions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: object
      responses:
        "201":
          description: Updated classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /ai/classifications/{id}/confirm:
    post:
      operationId: confirmClassification
      summary: Confirm proposed classification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: idempotency-key
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Confirmed classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /ai/classifications/{id}/override:
    post:
      operationId: overrideClassification
      summary: Manually override classification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subcategoryCode]
              properties:
                subcategoryCode:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Updated classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseClassificationRecord"
  /cases/{caseId}/messages:
    get:
      operationId: listMessages
      summary: List case messages
      parameters:
        - in: path
          name: caseId
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: addMessage
      summary: Add message to case chat
      parameters:
        - in: path
          name: caseId
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: idempotency-key
          required: false
          schema: {type: string, maxLength: 160}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /voice/transcripts:
    post:
      operationId: createTranscript
      summary: Create local/stub transcript job
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranscriptRequest"
      responses:
        "201":
          description: Transcript job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptJob"
  /voice/transcripts/audio:
    post:
      operationId: createAudioTranscript
      summary: Upload real audio and create transcript job
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                caseId:
                  type: string
                language:
                  type: string
                  enum: [ru, kk, en]
                text:
                  type: string
      responses:
        "201":
          description: Transcript job with stored audio metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptJob"
  /voice/transcripts/{id}:
    get:
      operationId: getTranscript
      summary: Get transcript job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Transcript job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptJob"
...
components:
  schemas:
...
    LegalCase:
      type: object
      required: [id, ownerUserId, title, problemText, category, confidence, status, readinessPercent, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
        profileId:
          type: string
        title:
          type: string
        problemText:
          type: string
        category:
          type: string
        subcategory:
          type: string
        confidence:
          type: number
        status:
          type: string
          enum: [consultation, clarification_required, collecting_documents, document_drafting, review_required]
        readinessPercent:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time
    CaseCategory:
      type: object
      required: [id, subcategory, title, procedure, description, criteria, keywords, officialSourceRefs]
      properties:
        id:
          type: string
        subcategory:
          type: string
        title:
          type: string
        procedure:
          type: string
          enum: [civil, administrative_public_law, administrative_offense, criminal, enforcement, special]
        description:
          type: string
        criteria:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        negativeKeywords:
          type: array
          items:
            type: string
        highRisk:
          type: boolean
        officialSourceRefs:
          type: array
          items:
            type: string
    LegalCategory:
      type: object
      required: [id, code, nameRu, nameKk, nameEn, active, highRisk, sortOrder, requiredFactSchema, requiredDocumentRules, clarificationQuestionTemplates, defaultLegalRoute, version]
      properties:
        id:
          type: string
        code:
          type: string
        parentId:
          type: string
        nameRu:
          type: string
        nameKk:
          type: string
        nameEn:
          type: string
        descriptionRu:
          type: string
        descriptionKk:
          type: string
        descriptionEn:
          type: string
        active:
          type: boolean
        highRisk:
          type: boolean
        sortOrder:
          type: integer
        requiredFactSchema:
          type: object
        requiredDocumentRules:
          type: object
        clarificationQuestionTemplates:
          type: array
          items:
            type: object
        defaultLegalRoute:
          type: string
        version:
          type: integer
    LegalCategoryChangePayload:
      type: object
      properties:
        code:
          type: string
        parentCode:
          type: string
        nameRu:
          type: string
        nameKk:
          type: string
        nameEn:
          type: string
        descriptionRu:
          type: string
        descriptionKk:
          type: string
        descriptionEn:
          type: string
        active:
          type: boolean
        highRisk:
          type: boolean
        sortOrder:
          type: integer
        requiredFactSchema:
          type: object
        requiredDocumentRules:
          type: object
        clarificationQuestionTemplates:
          type: array
          items:
            type: object
        defaultLegalRoute:
          type: string
          enum: [civil, administrative, enforcement, criminal_high_risk, manual_review]
    CreateLegalCategoryChangeRequest:
      type: object
      required: [action, categoryCode, payload]
      properties:
        action:
          type: string
          enum: [create, update]
        categoryCode:
          type: string
        payload:
          $ref: "#/components/schemas/LegalCategoryChangePayload"
        reason:
          type: string
    LegalCategoryChangeRequest:
      type: object
      required: [id, action, categoryCode, payload, status, requestedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          enum: [create, update]
        categoryCode:
          type: string
        payload:
          $ref: "#/components/schemas/LegalCategoryChangePayload"
        status:
          type: string
          enum: [pending, approved, rejected]
        reason:
          type: string
        requestedBy:
          type: string
        reviewedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
    CreateClassificationRequest:
      type: object
      required: [text]
      properties:
        caseId:
          type: string
        conversationId:
          type: string
        inputMessageId:
          type: string
        text:
          type: string
          minLength: 4
...
    CaseClassificationRecord:
      type: object
      required: [id, ownerUserId, result, userConfirmed, userOverridden, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
        caseId:
          type: string
        conversationId:
          type: string
        inputMessageId:
          type: string
        result:
          $ref: "#/components/schemas/StructuredClassificationResult"
        userConfirmed:
          type: boolean
        userOverridden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
    CreateCaseRequest:
      type: object
      required: [ownerUserId, problemText]
      properties:
        ownerUserId:
          type: string
        profileId:
          type: string
        classificationId:
          type: string
          format: uuid
          description: Confirmed classification with completed clarifications, preserved on the case
        problemText:
          type: string
          minLength: 10
    Message:
      type: object
      required: [id, caseId, role, text, createdAt]
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateMessageRequest:
      type: object
      required: [role, text]
      properties:
        role:
          type: string
          enum: [user]
        text:
          type: string
    CreateTranscriptRequest:
      type: object
      properties:
        caseId:
          type: string
        language:
          type: string
          enum: [ru, kk, en]
        audioRef:
          type: string
        text:
          type: string
    TranscriptJob:
      type: object
      required: [id, ownerUserId, status, language, transcript, lowConfidenceFragments, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        caseId:
          type: string
        status:
          type: string
          enum: [transcribing, classifying, retrieving_sources, validating, generating, review_required, ready, failed]
        language:
          type: string
          enum: [ru, kk, en]
        transcript:
          type: string
        lowConfidenceFragments:
          type: array
          items:
            type: string
        audioFileId:
          type: string
          format: uuid
        audioMimeType:
          type: string
        audioSizeBytes:
          type: integer
        audioSha256:
          type: string
        audioStorageKey:
          type: string
        progress:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...